refactor(renderer): extract launcher action model
